requirements.txt
google-cloud-secret-manager==2.5.0
google-cloud-storage
